% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_cb.R
\name{copy_cb}
\alias{copy_cb}
\title{Write a data frame to the clipboard.}
\usage{
copy_cb(x, sep = ";", col.names = T)
}
\arguments{
\item{x}{the object to be written, preferably a matrix or data frame.
    If not, it is attempted to coerce \code{x} to a data frame.}

\item{sep}{the field separator string.  Values within each row of
    \code{x} are separated by this string.}

\item{col.names}{either a logical value indicating whether the column
    names of \code{x} are to be written along with \code{x}, or a
    character vector of column names to be written.  See the section on
    \sQuote{CSV files} for the meaning of \code{col.names = NA}.}
}
\description{
This function is a wrapper around \code{\link{write.table}},
essentially setting its \code{file} parameter to \code{pipe("pbcopy")}.
}
\examples{
library(datasets)
data(mtcars)
copy_cb(mtcars)
copy_cb(mtcars[1:4], ',', F)
}
