% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dir.R
\name{create_dir}
\alias{create_dir}
\title{Create a directory including full path outside of the working directory}
\usage{
create_dir(dir, fullPath = F)
}
\arguments{
\item{dir}{A character string defining the path to the directory that should
be created}

\item{fullPath}{A boolean flag indicating whether the directory string
provided is a full path or not. If \code{FALSE}, \code{dir} is created in the user's
home directory as defined by \code{Sys.getenv("HOME")}.}
}
\description{
This function checks whether a directory in a given path exists and creates
it - including any missing directories in the path - if that is not the case.
}
\details{
This function is generally useful in the setup process of any (part of a)
project where the existence of a specific directory is dependent on the first
execution of a related part of your code base.
}
\examples{
create_dir('my/export/directory')

create_dir('~/my/export/directory', TRUE)

# using the get_file_path() function in jaytools
create_dir(get_file_path('my/export/directory'), TRUE)

}
