% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file_path.R
\name{get_file_path}
\alias{get_file_path}
\title{Create a character string for an absolute, fully resolved file path.}
\usage{
get_file_path(relativePath = "")
}
\arguments{
\item{relativePath}{Optional. A character string defining the relative path under the base path. See the Details section for more on the base path.}
}
\value{
A character string with the full expanded path (using \link{path.expand}.
}
\description{
This function constructs a full file path string for use in reading from or writing to the file system. It is always using a base path by default, and adds any additional relative path under it when one is passed to the function.
}
\details{
The base path defaults to the user directory (\code{~}), and can be defined with an environment variable \code{JT_LOCAL_PATH}.
}
\examples{
get_file_path('my/storage/path')
get_file_path('my/storage/path/for/this_file.csv')

Sys.setenv(JT_LOCAL_PATH = '~/my/custom/storage/path')
get_file_path('for/this_file.csv')
}
